Removed files to clear confusion
I had pullData.php in each test along with queryList.csv but these
should be on the local host, aka htdocs. So to prevent frustration of
modifying a file that is not actually used I deleted them a left this
comment. A sample queryList.csv has been provided for you in DB Scripts
folder along with the pullData.php..