reformatted filter.
now is slightly faster as its not making a copy each time and passing it
off. I am going to work on cleaner algorithm for doing the compressing
of the data. Im 95% sure I can do it in a way thats less lines and just
as / more efficient which makes it better for larger amounts of data